laptop_dashboard/requirements.txt
paho-mqtt>=2.0.0
matplotlib>=3.8.0